+++ main.cpp
@@ -5,7 +5,7 @@
 #include <memory>
 
 int main(int argc, char **argv) {
-	int N = 3;
+	size_t N = 3;
 	if(argc>=2)
 		N = atoi(argv[1]);
